+++ src/config/config.types.d.ts
@@ -41,6 +41,7 @@ interface YAMLConfig {
     imageEdits: string[];
     voices: string[];
     pronunciations: [[string, string]];
+    memoryCount: number;
   };
 
   twitter: {